requirements.txt
asgiref==3.7.2
Django==4.2.2
django-cors-headers==4.1.0
django-modeltranslation==0.18.10
djangorestframework==3.14.0
Pillow==9.5.0
python-decouple==3.8
pytz==2023.3
sqlparse==0.4.4
typing-extensions==4.6.3
